Abstract Worker functions
