+++ src/worker/cluster-worker.ts
@@ -1,3 +1,4 @@
+import type { Worker } from 'cluster'
 import { isMaster, worker } from 'cluster'
 import type { MessageValue } from '../utility-types'
 import { AbstractWorker } from './abstract-worker'
@@ -14,6 +15,7 @@ import type { WorkerOptions } from './worker-options'
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export class ClusterWorker<Data = any, Response = any> extends AbstractWorker<
+  Worker,
   Data,
   Response
 > {
@@ -21,10 +23,9 @@ export class ClusterWorker<Data = any, Response = any> extends AbstractWorker<
     super('worker-cluster-pool:pioardi', isMaster, fn, opts)
 
     worker.on('message', (value: MessageValue<Data>) => {
-      // console.log("cluster.on('message', value)", value)
       if (value?.data && value.id) {
         // here you will receive messages
-        // console.log('This is the main worker ' + isMaster)
+        // console.log('This is the main worker ' + isMain)
         if (this.async) {
           this.runInAsyncScope(this.runAsync.bind(this), this, fn, value)
         } else {
@@ -38,41 +39,15 @@ export class ClusterWorker<Data = any, Response = any> extends AbstractWorker<
     })
   }
 
-  protected checkAlive (): void {
-    if (Date.now() - this.lastTask > this.maxInactiveTime) {
-      worker.send({ kill: 1 })
-    }
+  protected getMainWorker (): Worker {
+    return worker
   }
 
-  protected run (
-    fn: (data?: Data) => Response,
-    value: MessageValue<Data>
-  ): void {
-    try {
-      const res = fn(value.data)
-      worker.send({ data: res, id: value.id })
-      this.lastTask = Date.now()
-    } catch (e) {
-      const err = e instanceof Error ? e.message : e
-      worker.send({ error: err, id: value.id })
-      this.lastTask = Date.now()
-    }
+  protected sendToMainWorker (message: MessageValue<Response>): void {
+    this.getMainWorker().send(message)
   }
 
-  protected runAsync (
-    fn: (data?: Data) => Promise<Response>,
-    value: MessageValue<Data>
-  ): void {
-    fn(value.data)
-      .then(res => {
-        worker.send({ data: res, id: value.id })
-        this.lastTask = Date.now()
-        return null
-      })
-      .catch(e => {
-        const err = e instanceof Error ? e.message : e
-        worker.send({ error: err, id: value.id })
-        this.lastTask = Date.now()
-      })
+  protected handleError (e: Error | string): string {
+    return e instanceof Error ? e.message : e
   }
 }
