+++ src/worker/thread-worker.ts
@@ -14,6 +14,7 @@ import type { WorkerOptions } from './worker-options'
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export class ThreadWorker<Data = any, Response = any> extends AbstractWorker<
+  MessagePort,
   Data,
   Response
 > {
@@ -25,7 +26,7 @@ export class ThreadWorker<Data = any, Response = any> extends AbstractWorker<
     parentPort?.on('message', (value: MessageValue<Data>) => {
       if (value?.data && value.id) {
         // here you will receive messages
-        // console.log('This is the main thread ' + isMainThread)
+        // console.log('This is the main worker ' + isMain)
         if (this.async) {
           this.runInAsyncScope(this.runAsync.bind(this), this, fn, value)
         } else {
@@ -36,46 +37,21 @@ export class ThreadWorker<Data = any, Response = any> extends AbstractWorker<
         // this will be received once
         this.parent = value.parent
       } else if (value.kill) {
-        // here is time to kill this thread, just clearing the interval
+        // here is time to kill this worker, just clearing the interval
         if (this.interval) clearInterval(this.interval)
         this.emitDestroy()
       }
     })
   }
 
-  protected checkAlive (): void {
-    if (Date.now() - this.lastTask > this.maxInactiveTime) {
-      this.parent?.postMessage({ kill: 1 })
+  protected getMainWorker (): MessagePort {
+    if (!this.parent) {
+      throw new Error('Parent was not set')
     }
+    return this.parent
   }
 
-  protected run (
-    fn: (data?: Data) => Response,
-    value: MessageValue<Data>
-  ): void {
-    try {
-      const res = fn(value.data)
-      this.parent?.postMessage({ data: res, id: value.id })
-      this.lastTask = Date.now()
-    } catch (e) {
-      this.parent?.postMessage({ error: e, id: value.id })
-      this.lastTask = Date.now()
-    }
-  }
-
-  protected runAsync (
-    fn: (data?: Data) => Promise<Response>,
-    value: MessageValue<Data>
-  ): void {
-    fn(value.data)
-      .then(res => {
-        this.parent?.postMessage({ data: res, id: value.id })
-        this.lastTask = Date.now()
-        return null
-      })
-      .catch(e => {
-        this.parent?.postMessage({ error: e, id: value.id })
-        this.lastTask = Date.now()
-      })
+  protected sendToMainWorker (message: MessageValue<Response>): void {
+    this.getMainWorker().postMessage(message)
   }
 }
